Split continues — dataset preparation & image processing moved to core/dataset/ .
Continues the split. Dataset preparation — analysis, config, image converter,
image prep, look consistency — is relocated into five modules under
core/dataset/. Pure move: all classes and functions are preserved; the main file
imports them when building the relevant tabs. Monolith: 3,372 → 1,456 lines.
Per-module refactoring is deferred.
